Create hospital visit audit table
